+++ src/hooks/worker.ts
@@ -6,12 +6,15 @@
 
 import { WorkerInfo } from '@playwright/test';
 import { fixtureParameterNames } from '../playwright/fixtureParameterNames';
-import { KeyValue, TestTypeCommon } from '../playwright/types';
+import { KeyValue, PlaywrightLocation, TestTypeCommon } from '../playwright/types';
 import { callWithTimeout } from '../utils';
+import { getLocationByOffset } from '../playwright/getLocationInFile';
+import { runStepWithLocation } from '../playwright/runStepWithLocation';
 
 export type WorkerHookType = 'beforeAll' | 'afterAll';
 
 type WorkerHookOptions = {
+  name?: string;
   timeout?: number;
 };
 
@@ -26,10 +29,11 @@ export type WorkerHook<Fixtures = object> = {
   type: WorkerHookType;
   options: WorkerHookOptions;
   fn: WorkerHookFn<Fixtures>;
-  // Since playwright-bdd v8 we run worker hooks via test.beforeAll / test.afterAll in each test file,
-  // so we need to know if hook was executed to avoid double execution every test file.
-  executed?: boolean;
+  location: PlaywrightLocation;
   customTest?: TestTypeCommon;
+  // Since playwright-bdd v8 we run worker hooks via test.beforeAll / test.afterAll in each test file,
+  // so we need to know if hook was executed to avoid double execution in every test file.
+  executed?: boolean;
 };
 
 /**
@@ -58,18 +62,24 @@ export function createBeforeAllAfterAll<Fixtures extends KeyValue>(
       type,
       options: getOptionsFromArgs(args) as WorkerHookOptions,
       fn: getFnFromArgs(args) as WorkerHook['fn'],
+      // offset = 3 b/c this call is 3 steps below the user's code
+      location: getLocationByOffset(3),
       customTest,
     });
   };
 }
 
 // eslint-disable-next-line visual/complexity
-export async function runWorkerHooks(type: WorkerHookType, fixtures: WorkerHookFixtures) {
+export async function runWorkerHooks(
+  test: TestTypeCommon,
+  type: WorkerHookType,
+  fixtures: WorkerHookFixtures,
+) {
   const hooksToRun = getWorkerHooksToRun(type);
   let error;
   for (const hook of hooksToRun) {
     try {
-      await runWorkerHook(hook, fixtures);
+      await runWorkerHook(test, hook, fixtures);
     } catch (e) {
       if (type === 'beforeAll') throw e;
       if (!error) error = e;
@@ -78,11 +88,12 @@ export async function runWorkerHooks(type: WorkerHookType, fixtures: WorkerHookF
   if (error) throw error;
 }
 
-async function runWorkerHook(hook: WorkerHook, fixtures: WorkerHookFixtures) {
+async function runWorkerHook(test: TestTypeCommon, hook: WorkerHook, fixtures: WorkerHookFixtures) {
   if (!hook.executed) {
     hook.executed = true;
-    const { timeout } = hook.options;
-    await callWithTimeout(() => hook.fn(fixtures), timeout, getTimeoutMessage(hook));
+    const hookFn = wrapHookFnWithTimeout(hook, fixtures);
+    const stepTitle = getHookStepTitle(hook);
+    await runStepWithLocation(test, stepTitle, hook.location, hookFn);
   }
 }
 
@@ -100,6 +111,22 @@ export function getWorkerHooksFixtureNames(hooks: WorkerHook[]) {
   });
 
   return [...fixtureNames];
+}
+
+/**
+ * Wraps hook fn with timeout.
+ */
+function wrapHookFnWithTimeout(hook: WorkerHook, fixtures: WorkerHookFixtures) {
+  const { timeout } = hook.options;
+
+  return async () => {
+    await callWithTimeout(
+      // call with null to avoid using 'this' inside worker hook
+      () => hook.fn.call(null, fixtures),
+      timeout,
+      getTimeoutMessage(hook),
+    );
+  };
 }
 
 function getOptionsFromArgs(args: unknown[]) {
@@ -123,4 +150,8 @@ function addHook(hook: WorkerHook) {
 function getTimeoutMessage(hook: WorkerHook) {
   const { timeout } = hook.options;
   return `${hook.type} hook timeout (${timeout} ms)`;
+}
+
+function getHookStepTitle(hook: WorkerHook) {
+  return hook.options.name || (hook.type === 'beforeAll' ? 'BeforeAll hook' : 'AfterAll hook');
 }