+++ test/reporter-data/features/steps.ts
@@ -2,7 +2,7 @@
 import { createBdd } from 'playwright-bdd';
 import { test } from './fixtures';
 
-const { Given, When, Before, BeforeAll } = createBdd(test);
+const { Given, When, Before, BeforeAll, AfterAll } = createBdd(test);
 
 Given('background step', async () => {});
 Given('I am on home page', async ({ myPage }) => {
@@ -12,3 +12,5 @@ When('Action {int}', () => {});
 Before({ name: 'hook 1' }, async () => {});
 Before(async () => {});
 BeforeAll(async () => {});
+BeforeAll({ name: 'named BeforeAll hook' }, async () => {});
+AfterAll({ name: 'named AfterAll hook' }, async () => {});