+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * support scoped step definitions ([#205](https://github.com/vitalets/playwright-bdd/issues/205))
 * new config option `missingSteps` to setup different behavior when step definitions are missing ([#158](https://github.com/vitalets/playwright-bdd/issues/158))
 * new config option `matchKeywords` to enable keyword matching when searching for step definitions ([#221](https://github.com/vitalets/playwright-bdd/issues/221))
+* add `name` option to `BeforeAll / AfterAll` hooks
 * make config option `quote` default to `"single"` to have less escapes in the generated files
 * config option `featuresRoot` serves as a default directory for both `features` and `steps`, if these options are not explicitly defined
 * feature: provide full localized step titles to Playwright HTML reporter ([#229](https://github.com/vitalets/playwright-bdd/issues/229), [#122](https://github.com/vitalets/playwright-bdd/issues/122))